+++ Prog0/src/sdns/serialization/NS.java
@@ -80,7 +80,7 @@ public class NS extends ResourceRecord {
      * @return type value
      */
     @Override
-    public long getTypeValue() { return NS_TYPE_VALUE; }
+    public int getTypeValue() { return NS_TYPE_VALUE; }
 
     /**
      * Returns a byte array of the rdata for this object.  For internal use only.
